requirements.txt
Django==4.2.4
djangorestframework==3.15.2
django-filter==23.1
drf-yasg
PyOTP
python-decouple
